Lock version of openssl-sys to known working version
The update breaks building of openssl-sys on non-linux targets, needed for the
internal releases. Will remain until it becomes clearer if this is a bug
in the crate or our build infra. (At the very least, we are not the only ones with
this issue. See: https://github.com/rust-openssl/rust-openssl/issues/2514 )

Signed-off-by: Jonatan Waern <jonatan.waern@intel.com>